backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.3.0
python-dotenv==1.0.0
pillow==10.1.0
aiofiles==23.2.1
jinja2==3.1.2
redis==5.0.1
celery==5.3.4
python-magic==0.4.27
psutil==5.9.6
pydantic==2.5.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4